Extract search, fetch, and cache expiry into custom hooks
