+++ src/App.jsx
@@ -1,9 +1,8 @@
 import React, { useState, useEffect, useContext } from "react";
 
-import { baseURL } from './data';
-
-import { getSearchURL } from "./utilities";
-
+import { useExpireListingIDs } from "./hooks/useExpireListingIDs";
+import { useFetchListings } from "./hooks/useFetchListings";
+import { useSearchHandler } from "./hooks/useSearchHandler";
 import { ListingsContext } from ".";
 
 import ListingsContainer from "./components/listings/ListingsContainer";
@@ -19,22 +18,10 @@ const App = () => {
   const [search, setSearch] = useState(false);
   const [searchQuery, setSearchQuery] = useState([]);
   const { listingIDs, setListingIDs } = useContext(ListingsContext);
-
-  // Delete previous search results if more than 2 days old
-  useEffect(() => {
-    const currentTime = new Date();
-    var dateOffset = (24 * 60 * 60 *1000) * 2;
-    const date48HoursAgo = currentTime.getTime() - dateOffset;
-
-    const lastSearchTime = localStorage.getItem("lastSearchTime")
-    if (!lastSearchTime) return;
-
-    if (Number(lastSearchTime) < date48HoursAgo) {
-      console.log("Search results too old, removing listingIDs")
-      localStorage.removeItem("lastSearchTime")
-      localStorage.removeItem("listingIDs")
-    }
-  }, []);
+  
+  useExpireListingIDs();
+  useFetchListings(queryURL, setListingIDs, setNoListingsFound, setQueryURL, setSearch);
+  useSearchHandler(search, searchQuery, agentChoices, setQueryURL, setSearch);
 
   useEffect(() => {
     if (!noListingsFound && !listingIDs?.length) {
@@ -43,43 +30,7 @@ const App = () => {
   }, [listingIDs?.length, noListingsFound, setListingIDs]);
 
   useEffect(() => {
-    if (typeof queryURL === "string") {
-      fetch(baseURL + queryURL)
-        .then(res => res.json())
-        .then(data => {
-          const hiddenListings = JSON.parse(localStorage.getItem("hiddenListings")) || [];
-          // Sorting listings by most recently added means they are effectively sorted by agent, so to display them in a way that is somewhat randomized but will always return the same results in the same order, sort them using a combination of their listingID and house size
-          const sortedListings = data?.length 
-            ? data
-                .filter(listing => !hiddenListings.includes(listing.listingID))
-                .sort((a, b) => a.listingID * (a.size || 1) < b.listingID * (b.size || 1) ? 1 : -1) 
-            : [];
-          setListingIDs(sortedListings);
-          // save array of shortened listings to local storage
-          localStorage.setItem("listingIDs", JSON.stringify(sortedListings));
-          // Set lastSearchTime in order to ensure listings expire after 48 hours
-          const timeNow = new Date();
-          localStorage.setItem("lastSearchTime", JSON.stringify(timeNow.getTime()));
-          setNoListingsFound(!sortedListings?.length);
-          setQueryURL(null);
-          setSearch(false);
-        })
-        .catch(err => console.error(err));
-    }
-  }, [queryURL, setListingIDs]);
-
-  useEffect(() => {
-    if (search) {
-      const searchURL = getSearchURL(searchQuery, agentChoices);
-      setQueryURL(searchURL);
-      setSearch(false);
-    }
-  }, [search, searchQuery]);
-
-  useEffect(() => {
-    document.body.style.overflow = loadingListings 
-      ? "hidden"
-      : "auto"
+    document.body.style.overflow = loadingListings ? "hidden" : "auto";
   }, [loadingListings]);
 
   return (
